requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-parler==2.3
django-rosetta==0.9.8
flake8==4.0.1
idna==3.3
mccabe==0.6.1
polib==1.1.1
pycodestyle==2.8.0
pyflakes==2.4.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
